requirements.txt
streamlit
pandas
psycopg2-binary
python-dotenv
langchain
langchain-openai
langgraph
pgvector
sqlalchemy